add golang-lib

+++ src/template/golang_lib.ts
@@ -1,17 +1,19 @@
 import InterfaceTemplate from './interface_template'
 import ShellHelper from '@pefish/js-helper-shell'
 
-export default class TsLib implements InterfaceTemplate {
+export default class GolangLib implements InterfaceTemplate {
   getTemplateRepoUrl(): string {
-    return `git@github.com:pefish/create-typescript-lib-template.git`
+    return `git@github.com:pefish/create-golang-lib-template.git`
   }
 
   do(shellHelper: ShellHelper, projectName: string, desc: string, opts: {[x: string]: any}): void {
+    if (!opts.packageName) {
+      throw new Error(`packageName must be setted`)
+    }
     shellHelper.execSync(`git clone ${this.getTemplateRepoUrl()} --single-branch -v -b master --depth 1 ${projectName}`)
     shellHelper.cd(projectName)
     shellHelper.execSync(`rm -rf .git`)
-    shellHelper.execSync(`cat package.json | sed "s/template_name/${projectName}/g" > temp.json && rm -rf package.json && mv temp.json package.json`)
-    shellHelper.execSync(`cat package.json | sed "s/template_description/${desc}/g" > temp.json && rm -rf package.json && mv temp.json package.json`)
+    shellHelper.execSync(`cat go.mod | sed "s/create_golang_lib_template/${opts.packageName}/g" > temp && rm -rf go.mod && mv temp go.mod`)
     shellHelper.execSync(`yarn`)
   }
